+++ PracticeProblems/nQueens.h
@@ -11,14 +11,14 @@ vector<int> I used a single variable and bit-shifting to solve the problem.
 using namespace std;
 class NQueens {
 public:
-	vector<vector<string>> solveNQueens(int n);
+	vector<vector<string>> solveNQueens(uint64_t n);
 private:
 	void helper(vector<vector<string>> &answers,
 		vector<string> &board,
-		vector<int> &queens,
+		vector<uint64_t> &queens,
 		uint64_t &colMask,
 		uint64_t &deg45Mask,
 		uint64_t &deg135Mask,
-		int &n,
-		int row);
+		uint64_t &n,
+		uint64_t row);
 };
